% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abronia}
\alias{abronia}
\title{Occurrence records of Abronia graminea}
\format{
A data frame with 106 rows and 5 variables:
\describe{
\item{species}{Scientific name of the species}
\item{decimalLongitude}{Longitude}
\item{decimalLatitude}{Latitude}
\item{year}{Observation year}
\item{gbif_doi}{DOI id for citing the dataset}
...
}
}
\source{
GBIF.org (22 February 2022) GBIF Occurrence
Download \doi{10.15468/dl.teyjm9}
}
\usage{
abronia
}
\description{
A dataset containing occurrence records for Abronia graminea.
The data was downloaded from GBIF (GBIF, 2022).
}
\keyword{datasets}
