% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells2samp.R
\name{cells2samp}
\alias{cells2samp}
\title{Helper function to generate cell IDs to be used for choosing  the
environmental background data}
\usage{
cells2samp(
  data,
  longitude,
  latitude,
  cell_ids = NULL,
  buffer_ngbs = 2,
  raster_mask,
  process_ngbs_by = 100,
  n_bg = 50000,
  progress = TRUE
)
}
\arguments{
\item{data}{A data.frame with longitude and latitude data}

\item{longitude}{A character vector of the column name of longitude.}

\item{latitude}{A character vector of the column name of latitude.}

\item{cell_ids}{A numeric vector with pixel ids. The default values NULL.
Use this parameter if you have obtained this information using the function
cellFromXY.}

\item{buffer_ngbs}{Number of pixel neighbors around occurrences to be used
to build the buffer.}

\item{raster_mask}{An object of class RasterLayer that will be used to
obtain pixel IDs.}

\item{process_ngbs_by}{Numeric. Estimates neighbor cells each x cells. This
is for memory management.}

\item{n_bg}{Number of background pixels.}

\item{progress}{Logical. Show computation progress}
}
\description{
Returns pixel IDs to be sample for generating
environmental background form modeling layers.
}
