% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors}
\alias{colors}
\title{Colors for plotting}
\format{
An object of class \code{character} of length 40.
}
\usage{
colors
}
\description{
A string vector of colors for plotting the vignette example.
}
\keyword{datasets}
