% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_finder.R
\name{correlation_finder}
\alias{correlation_finder}
\title{Function to find out strong correlations in a correlation matrix}
\usage{
correlation_finder(
  environmental_data,
  method = "spearman",
  threshold,
  verbose = TRUE
)
}
\arguments{
\item{environmental_data}{A matrix or a data.frame of environmental data}

\item{method}{A method to estimate correlation matrix. Posible options are "spearman", "pearson" or
"kendall".}

\item{threshold}{Threshold value from which it is considered that the correlation is high.}

\item{verbose}{Verbose output.}
}
\value{
Returns a vector with variable names that can summarize the information.
}
\description{
The function finds out which variables have strong
correlations according to a correlation threshold. The output
returns a list of variables names that can summarize the information
and removes the variables that are redundant.
}
