% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_center.R
\name{cov_center}
\alias{cov_center}
\title{Function to compute the Minimum Volume covariance Matrix of an ellipsoid niche model.}
\usage{
cov_center(data, mve = TRUE, level, vars = NULL)
}
\arguments{
\item{data}{A data.frame or a matrix with the numeric values of the variables
that will be used to model the niche.}

\item{mve}{A logical value. If TRUE a minimum volume ellipsoid will be computed using
the function \code{\link[MASS]{cov.mve}} of the \pkg{MASS} package. If False the covariance matrix of the input data will be used.}

\item{level}{A numerical value specifying the proportion of the data to be
used to compute the ellipsoid.}

\item{vars}{A numeric or a string vector specifying the columns indexes/names of the variables of the input data which will be used to fit the ellipsoid model. If NULL the user will be asked to enter the indexes.
interactively}
}
\value{
Returns a list containing the centroid of the ellipsoid, the covariance matrix based on
the input data, ellipsoid volume, semi-axis length and axis coordinates.
}
\description{
Function to compute the covariance matrix, the niche centroid and volume of an
ellipsoid model. It uses the values of the niche variables of the ocurrences points.
}
\examples{
\donttest{
library(tenm)
data("abronia")
tempora_layers_dir <- system.file("extdata/bio",package = "tenm")
abt <- tenm::sp_temporal_data(occs = abronia,
                              longitude = "decimalLongitude",
                              latitude = "decimalLatitude",
                              sp_date_var = "year",
                              occ_date_format="y",
                              layers_date_format= "y",
                              layers_by_date_dir = tempora_layers_dir,
                              layers_ext="*.tif$")
abtc <- tenm::clean_dup_by_date(abt,threshold = 10/60)
future::plan("multisession",workers=10)
abex <- tenm::ex_by_date(abtc,train_prop=0.7)
future::plan("sequential")
varcorrs <- tenm::correlation_finder(environmental_data = abex$env_data[,-ncol(abex$env_data)],
                                     method = "spearman",
                                     threshold = 0.8,
                                     verbose = FALSE)
mod <- tenm::cov_center(data = abex$env_data,
                        mve = TRUE,
                        level = 0.975,
                        vars = c("bio_05","bio_06","bio_12"))
# Print model parameters
print(mod)
}
}
