% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_projection.R
\name{ellipsoid_projection}
\alias{ellipsoid_projection}
\title{ellipsoid_projection: function to fit an ellipsoid model}
\usage{
ellipsoid_projection(
  envlayers,
  centroid,
  covar,
  level = 0.95,
  plot = T,
  size,
  xlab1 = "niche var 1",
  ylab1 = "niche var 2",
  zlab1 = "S",
  alpha = 0.1,
  ...
)
}
\arguments{
\item{envlayers}{A SpatRaster object of the niche variables.}

\item{centroid}{A vector with the values of the centers of the ellipsoid
(see \code{\link[tenm]{cov_center}}).}

\item{covar}{The shape matrix (covariance) of the ellipsoid
(see \code{\link[tenm]{cov_center}}).}

\item{level}{The proportion of points  to be included inside the ellipsoid}

\item{plot}{Logical If True a plot of the niche will be shown.}

\item{size}{The size of the points of the niche plot.}

\item{xlab1}{For x label for 2-dimensional histogram}

\item{ylab1}{For y label for 2-dimensional histogram}

\item{zlab1}{For z label for 2-dimensional histogram}

\item{alpha}{Control the transparency of the 3-dimensional ellipsoid}

\item{...}{Arguments passed to \code{\link[rgl]{plot3d}} function from rgl}
}
\value{
Returns a list containing a data.frame with the suitability values;
a suitability raster;
a data.frame with the mahalanobis and euclidean distances to the centroid.
}
\description{
Function to fit an ellipsoid model using the shape matrix
(covariance matrix) of the niche variables. This is an optimized version of
the ellipsoidfit function of ntbox; the difference is that it
does not give the table of distances to niche centroid.
}
