% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaras.R
\name{metaras}
\alias{metaras}
\title{Helper function to obtain layer name from a}
\usage{
metaras(r)
}
\arguments{
\item{r}{An object of class RasterLayer}
}
\description{
Returns a character vector with the name of the raster layer
}
\examples{
tempora_layers_dir <- system.file("extdata/bio",package = "tenm")
p1 <- list.files(tempora_layers_dir,full.names=TRUE,
                 pattern=".tif$",recursive=TRUE)[1]
r1 <- raster::raster(p1)
print(metaras(r1))
}
