% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_methods.R
\docType{methods}
\name{predict,sp.temporal.selection-method}
\alias{predict,sp.temporal.selection-method}
\alias{predict}
\title{Predict potential distribution for \pkg{tenm} regarding to a specific time
period of environmental conditions or averages.}
\usage{
\S4method{predict}{sp.temporal.selection}(
  object,
  model_variables = NULL,
  layers = NULL,
  layers_path = NULL,
  layers_ext = NULL,
  mve = TRUE,
  level = 0.975,
  ...
)
}
\arguments{
\item{object}{An object of class sp.temporal.selection}

\item{model_variables}{A vector with variable names}

\item{layers}{A SpatRaster object or a list where each element is a SpatRaster.}

\item{layers_path}{Path to layers}

\item{layers_ext}{Layers extension}

\item{mve}{If the projection will use the minimum volume ellipsoid algorithm}

\item{level}{Proportion of data used to fit the minimum volume ellipsoid}

\item{...}{Additional parameters passed to
\code{\link[tenm]{ellipsoid_projection}}}
}
\description{
Predict potential distribution for \pkg{tenm} regarding to a specific time
period of environmental conditions or averages.
}
\details{
Note that the SpatRaster in layers parameter should have the same
number of elements (layers) than model_variables. The predict method will
assume that variables in each SpatRaster are the same as the ones in
model_variables.
}
