% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{sp.temporal.bg-class}
\alias{sp.temporal.bg-class}
\title{S3 classes to organize data and results of \code{tenm} objects}
\value{
An object of class 'sp.temporal.bg'. The object inherits information
from objects of classes 'sp.temporal.modeling' and 'sp.temporal.env'. This
class adds environmental background information.
}
\description{
S3 classes to organize data and results of \code{tenm} objects
}
\author{
Luis Osorio-Olvera
}
