% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{sp.temporal.env-class}
\alias{sp.temporal.env-class}
\title{S3 classes to organize data and results of \code{tenm} objects}
\value{
An object of class 'sp.temporal.env' inheriting information from
'sp.temporal.env'. This object adds a data.frame of environmental values
associated with occurrence data.
}
\description{
S3 classes to organize data and results of \code{tenm} objects
}
\author{
Luis Osorio-Olvera
}
