% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{sp.temporal.modeling-class}
\alias{sp.temporal.modeling-class}
\title{S3 classes to organize data and results of \code{tenm} objects}
\value{
This object is a list comprising four elements: a) A data.frame
containing occurrence records and layer information. b) A character vector
specifying variable names. c) A character vector indicating the names of
longitude and latitude variables. d) A character vector denoting the
layers extension.
}
\description{
S3 classes to organize data and results of \code{tenm} objects
}
\author{
Luis Osorio-Olvera
}
