% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{sp.temporal.selection-class}
\alias{sp.temporal.selection-class}
\title{S3 classes to organize data and results of \code{tenmp} objects}
\description{
S3 classes to organize data and results of \code{tenmp} objects
}
\author{
Luis Osorio-Olvera
}
