% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suit_1970_2000}
\alias{suit_1970_2000}
\title{Potential distribution of abronia graminea}
\format{
An object of class \code{RasterLayer} of dimension 23 x 20 x 1.
}
\usage{
suit_1970_2000
}
\description{
A dataset containing a niche model for Abronia graminea.
The model was generated using the time-specific modeling approach implemented
in the \emph{tenm} package.
}
\keyword{datasets}
