% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_by_date.R
\name{bg_by_date}
\alias{bg_by_date}
\title{Function to obtain environmental background organized by date}
\usage{
bg_by_date(
  this_species,
  buffer_ngbs = NULL,
  buffer_distance = 1000,
  n_bg = 50000,
  process_ngbs_by = 100
)
}
\arguments{
\item{this_species}{Species Temporal Environmental Data (sp.temporal.env)
object from \code{\link[tenm]{ex_by_date}}.}

\item{buffer_ngbs}{Number of pixel neighbors used to build the buffer.}

\item{buffer_distance}{Distance used to create a buffer in which background
data will be taken.}

\item{n_bg}{Number of background points.}

\item{process_ngbs_by}{Numeric. Estimates neighbor cells each x cells. This
is for memory management.}
}
\value{
Returns an object of class sp.temporal.bg which is a list that
contains a data.frame with longitude, latitude, year, layer_date, layer_path,
cell_ids_year and environmental information.
}
\description{
Function to retrieve background data from occurrence records.
The background data is organized as a function of the dated
environmental data
}
\details{
The buffer is built around the occurrences using a neighborhood
distance.
}
\examples{
\dontrun{
library(tenm)
data("abronia")
tempora_layers_dir <- system.file("extdata/bio",package = "tenm")
abt <- tenm::sp_temporal_data(occs = abronia,
                              longitude = "decimalLongitude",
                              latitude = "decimalLatitude",
                              sp_date_var = "year",
                              occ_date_format="y",
                              layers_date_format= "y",
                              layers_by_date_dir = tempora_layers_dir,
                              layers_ext="*.tif$")
abtc <- tenm::clean_dup_by_date(abt,threshold = 10/60)
#This code is for running in parallel
future::plan("multisession",workers=2)
abex <- tenm::ex_by_date(this_species = abtc,train_prop=0.7)
abbg <- tenm::bg_by_date(this_species = abex,
                         buffer_ngbs=10,n_bg=50000)
future::plan("sequential")
}
}
