% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dup.R
\name{clean_dup}
\alias{clean_dup}
\title{Function to clean duplicated longitude and latitude data}
\usage{
clean_dup(
  data,
  longitude,
  latitude,
  threshold = 0,
  by_mask = FALSE,
  raster_mask = NULL,
  n_ngbs = 0
)
}
\arguments{
\item{data}{A data.frame with longitude and latitude data}

\item{longitude}{A character vector of the column name of longitude.}

\item{latitude}{A character vector of the column name of latitude.}

\item{threshold}{A numeric value representing the euclidean distance between coordinates
to be considered as a duplicate.}

\item{by_mask}{Logical. If TRUE the elimination of duplicates will be done using a raster layer as a mask;
If False the elimination of duplicates will be done by the distance threshold.}

\item{raster_mask}{An object of class RasterLayer that will be used to clean duplicates that are present in the same ID pixel.}

\item{n_ngbs}{DOCUMENTAR!!!! Number of pixel neighbors. Remove duplicates depending on how many pixels range you want, 1 is for eliminate duplicates in the same
pixel, 2 is a neighborhood of 3 for 3 pixels, 3 is an 5 for 5 vicinity and so on depending on how much area you want to cover.}
}
\value{
Returns a data.frame with coordinate data from species
}
\description{
Clean duplicated longitude and latitude data using threshold distance. In a process that we call "cleaning of duplicated", this function has the main propose of eliminating data occurrence points spliced, through the information of longitude and latitude,
identifying the redundant environmental data in a given distance threshold.That means the possibility of be able to debug large occurrence point clouds
that could cause an overestimation of the future model. This function also allows you to eliminate duplicate occurrence points per pixel or in a given pixel neighborhood.
}
\examples{
data(abronia)
data(suit_1970_2000)
# Clean duplicates without raster mask (just by distance threshold)
# First check the number of occurrence records
print(nrow(abronia))
# Clean duplicated records using a distance of ~ 18 km (0.1666667 grades)
ab_1 <- tenm::clean_dup(data =abronia,
                        longitude = "decimalLongitude",
                        latitude = "decimalLatitude",
                        threshold = raster::res(suit_1970_2000),
                        by_mask = FALSE,
                        raster_mask = NULL)
# Check number of records
print(nrow(ab_1))
# Clean duplicates using a raster mask
ab_2 <- tenm::clean_dup(data =abronia,
                        longitude = "decimalLongitude",
                        latitude = "decimalLatitude",
                        threshold = raster::res(suit_1970_2000),
                        by_mask = TRUE,
                        raster_mask = suit_1970_2000,
                        n_ngbs = 0)
# Check number of records
print(nrow(ab_2))
}
