% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dup_by_date.R
\name{clean_dup_by_date}
\alias{clean_dup_by_date}
\title{Function to clean duplicated coordinates data}
\usage{
clean_dup_by_date(
  this_species,
  threshold,
  by_mask = FALSE,
  raster_mask = NULL,
  n_ngbs = 0
)
}
\arguments{
\item{this_species}{Species Temporal Data object see \code{\link[tenm]{sp_temporal_data}}.}

\item{threshold}{A numeric value representing the distance between coordinates
to be considered as a duplicate.}

\item{by_mask}{Logical. If TRUE the elimination of duplicates will be done using a raster layer as a mask;
If False the elimination of duplicates will be done by the distance threshold.}

\item{raster_mask}{An object of class RasterLayer that will be used to clean duplicates that are present in the same ID pixel.}

\item{n_ngbs}{Number of pixel neighbors. Remove duplicates depending on how many pixels range you want,
"0" is use in order to remove all the duplicated occurrence points present in a single pixel (1x1), "1" is to eliminate duplicates in a 1 pixel-long
neighborhood of a 9 pixel area (3x3), "2" correspond to a neighborhood of 2 adjacent pixels of a 25 pixel area (5x5) and so
on depending on how much area you want to cover.}
}
\description{
Clean up duplicated longitude and latitude data by year using a threshold distance. The main propose of this function is eliminate data occurrence points spliced in each different year or date
by a given distance threshold. This function also allows you to clean duplicate data by pixel at a given resolution of a raster mask.
}
\details{
This function is build on the basis of \code{\link[tenm]{clean_dup}}.
See the help of the function for more examples
}
\examples{
library(tenm)
data("abronia")
data(suit_1970_2000)
tempora_layers_dir <- system.file("extdata/bio",package = "tenm")
# Clean duplicates without raster mask (just by distance threshold)
abt <- tenm::sp_temporal_data(occs = abronia,
                              longitude = "decimalLongitude",
                              latitude = "decimalLatitude",
                              sp_date_var = "year",
                              occ_date_format="y",
                              layers_date_format= "y",
                              layers_by_date_dir = tempora_layers_dir,
                              layers_ext="*.tif$")
abtc1 <- tenm::clean_dup_by_date(abt,threshold = raster::res(suit_1970_2000))
# Check number of records
print(nrow(abtc1$temporal_df))
# Clean duplicates using a raster mask
abtc2 <- tenm::clean_dup_by_date(this_species = abt,
                                by_mask = TRUE,
                                raster_mask = suit_1970_2000,
                                n_ngbs = 0)
# Check number of records
print(nrow(abtc2$temporal_df))

}
