% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_finder.R
\name{correlation_finder}
\alias{correlation_finder}
\title{Function to find out strong correlations in a correlation matrix}
\usage{
correlation_finder(
  environmental_data,
  method = "spearman",
  threshold,
  verbose = TRUE
)
}
\arguments{
\item{environmental_data}{A matrix or a data.frame of environmental data}

\item{method}{A method to estimate correlation matrix. Possible options are
"spearman", "pearson" or "kendall".}

\item{threshold}{Correlation value used to filter variables.}

\item{verbose}{Verbose output.}
}
\value{
Returns a list of two elements: the first is a vector with the names
of not correlated variables; the second is a list with the correlation values
of all variables.
}
\description{
The function finds out which variables have strong
correlations according to a correlation threshold.
}
\examples{
\donttest{

library(tenm)
data("abronia")
tempora_layers_dir <- system.file("extdata/bio",package = "tenm")
abt <- tenm::sp_temporal_data(occs = abronia,
                              longitude = "decimalLongitude",
                              latitude = "decimalLatitude",
                              sp_date_var = "year",
                              occ_date_format="y",
                              layers_date_format= "y",
                              layers_by_date_dir = tempora_layers_dir,
                              layers_ext="*.tif$")
abtc <- tenm::clean_dup_by_date(abt,threshold = 10/60)
future::plan("multisession",workers=2)
abex <- tenm::ex_by_date(abtc,train_prop=0.7)
future::plan("sequential")
envdata <- abex$env_data[,-ncol(abex$env_data)]
ecors <- tenm::correlation_finder(environmental_data =envdata,
                                  method="spearman",
                                  threshold = 0.7 )
}

}
